EES-4497 Move test data in-line on request
