+++ src/explore-education-statistics-common/src/modules/methodology/components/__tests__/MethodologyHelpAndSupportSection.test.tsx
@@ -2,9 +2,18 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import * as ContactUsSection from '@common/modules/find-statistics/components/ContactUsSection';
 import MethodologyHelpAndSupportSection from '@common/modules/methodology/components/MethodologyHelpAndSupportSection';
-import testPublicationSummary from '@common/modules/methodology/components/__tests__/__data__/test-data';
+import { PublicationSummary } from '@common/services/publicationService';
+import mockContact from '@common/modules/find-statistics/components/__tests__/__data__/test-data';
 
 jest.mock('@common/modules/find-statistics/components/ContactUsSection');
+
+const testPublicationSummary: PublicationSummary = {
+  id: 'Mock Publication Id',
+  slug: 'Mock Publication Slug',
+  title: 'Mock Publication Title',
+  owner: false,
+  contact: mockContact,
+};
 
 describe('MethodologyHelpAndSupportSection', () => {
   beforeEach(() => {
@@ -19,13 +28,15 @@ describe('MethodologyHelpAndSupportSection', () => {
     );
   });
 
-  test('that it renders', () => {
+  test('renders', () => {
     expect(
       screen.getByRole('heading', { name: /Help and support/ }),
     ).toBeVisible();
   });
 
-  test('that it renders a Contact Us Section', () => {
-    expect(screen.getByText('This is a mocked ContactUsSection')).toBeVisible();
+  test('renders a Contact Us Section', () => {
+    expect(
+      screen.getByText('This is a mocked ContactUsSection'),
+    ).toBeInTheDocument();
   });
 });
